requirements.txt
streamlit==1.22.0
pandas==1.5.3
torch==2.0.1
transformers==4.31.0
spacy==3.5.1
pandas==1.5.3
scikit-learn==1.2.2
pymupdf==1.21.1
spacy==3.5.1
pandas==1.5.3
scikit-learn==1.2.2
pymupdf==1.21.1
torch==2.0.1
transformers==4.31.0
python-Levenshtein==0.12.2
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
dataclasses-json==0.5.14
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
langchain==0.0.271
langsmith==0.0.79
marshmallow==3.20.2
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.8
numpy==1.26.3
packaging==23.2
pydantic==2.5.3
pydantic_core==2.14.6
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
PyYAML==6.0.1
requests==2.31.0
SQLAlchemy==2.0.25
tenacity==8.2.3
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
